Extract numbers from a string.

Extract the numbers from a string, where decimals, scientific notation and commas (as separators, not as an alternative to the decimal point) are optionally allowed.

str_extract_numbers(string, leave_as_string = FALSE, decimals = FALSE,
  leading_decimals = decimals, negs = FALSE, sci = FALSE,
  commas = FALSE)

str_first_number(string, leave_as_string = FALSE, decimals = FALSE,
  leading_decimals = decimals, negs = FALSE, sci = FALSE,
  commas = FALSE)

str_last_number(string, leave_as_string = FALSE, decimals = FALSE,
  leading_decimals = decimals, negs = FALSE, sci = FALSE,
  commas = FALSE)

Arguments

string
A string.

leave_as_string
Do you want to return the number as a string (TRUE) or as numeric (FALSE, the default)?

decimals
Do you want to include the possibility of decimal numbers (TRUE) or not (FALSE, the default).

leading_decimals
Do you want to allow a leading decimal point to be the start of a number?

negs
Do you want to allow negative numbers? Note that double negatives are not handled here (see the examples).

sci
Make the search aware of scientific notation e.g. 2e3 is the same as 2000.

commas
Allow comma separators in numbers (i.e. interpret 1,100 as a single number (one thousand one hundred) rather than two numbers (one and one hundred)).

Value

For str_extract_numbers and str_extract_non_numerics, a list of numeric or character vectors, one list element for each element of string. For str_nth_number and nth_non_numeric, a vector the same length as string (as in length(string), not nchar(string)).

Details

If any part of a string contains an ambiguous number (e.g. 1.2.3 would be ambiguous if decimals = TRUE (but not otherwise)), the value returned for that string will be NA and a warning will be issued.

With scientific notation, it is assumed that the exponent is not a decimal number e.g. 2e2.4 is unacceptable. Commas, however, are acceptable in the exponent, so 2e1,100 is fine and equal to 2e1100 if the option to allow commas in numbers has been turned on.

Numbers outside the double precision floating point range (i.e. with absolute value greater than 1.797693e+308) are read as Inf (or -Inf if they begin with a minus sign). This is what base::as.numeric() does.

Examples

strings <- c(
  "abc123def456", "abc-0.12def.345", "abc.12e4def34.5e9",
  "abc1,100def1,230.5", "abc1,100e3,215def4e1,000"
)
str_extract_numbers(strings)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1]   0  12 345
#> 
#> [[3]]
#> [1] 12  4 34  5  9
#> 
#> [[4]]
#> [1]   1 100   1 230   5
#> 
#> [[5]]
#> [1]   1 100   3 215   4   1   0
#> 

str_extract_numbers(strings, decimals = TRUE)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1] 0.120 0.345
#> 
#> [[3]]
#> [1]  0.12  4.00 34.50  9.00
#> 
#> [[4]]
#> [1]   1.0 100.0   1.0 230.5
#> 
#> [[5]]
#> [1]   1 100   3 215   4   1   0
#> 

str_extract_numbers(strings, decimals = TRUE, leading_decimals = TRUE)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1] 0.120 0.345
#> 
#> [[3]]
#> [1]  0.12  4.00 34.50  9.00
#> 
#> [[4]]
#> [1]   1.0 100.0   1.0 230.5
#> 
#> [[5]]
#> [1]   1 100   3 215   4   1   0
#> 

str_extract_numbers(strings, commas = TRUE)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1]   0  12 345
#> 
#> [[3]]
#> [1] 12  4 34  5  9
#> 
#> [[4]]
#> [1] 1100 1230    5
#> 
#> [[5]]
#> [1] 1100 3215    4 1000
#> 

str_extract_numbers(strings,
  decimals = TRUE, leading_decimals = TRUE,
  sci = TRUE
)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1] 0.120 0.345
#> 
#> [[3]]
#> [1] 1.20e+03 3.45e+10
#> 
#> [[4]]
#> [1]   1.0 100.0   1.0 230.5
#> 
#> [[5]]
#> [1]      1 100000    215     40      0
#> 

str_extract_numbers(strings,
  decimals = TRUE, leading_decimals = TRUE,
  sci = TRUE, commas = TRUE, negs = TRUE
)

#> [[1]]
#> [1] 123 456
#> 
#> [[2]]
#> [1] -0.120  0.345
#> 
#> [[3]]
#> [1] 1.20e+03 3.45e+10
#> 
#> [[4]]
#> [1] 1100.0 1230.5
#> 
#> [[5]]
#> [1] Inf Inf
#> 

str_extract_numbers(strings,
  decimals = TRUE, leading_decimals = FALSE,
  sci = FALSE, commas = TRUE, leave_as_string = TRUE
)

#> [[1]]
#> [1] "123" "456"
#> 
#> [[2]]
#> [1] "0.12" "345" 
#> 
#> [[3]]
#> [1] "12"   "4"    "34.5" "9"   
#> 
#> [[4]]
#> [1] "1,100"   "1,230.5"
#> 
#> [[5]]
#> [1] "1,100" "3,215" "4"     "1,000"
#>