Locate the indices of the \(n\)th instance of a pattern.

The \(n\)th instance of an pattern will cover a series of character indices. These functions tell you which indices those are.

str_locate_nth(string, pattern, n)

str_locate_first(string, pattern)

str_locate_last(string, pattern)

Arguments

string
A character vector. These functions are vectorized over this argument.

pattern
A character vector. Pattern(s) specified like the pattern(s) in the stringr package (e.g. look at stringr::str_locate()). If this has length >1 its length must be the same as that of string.

n
Then \(n\) for the \(n\)th instance of the pattern.

Value

A two-column matrix. The \(i\)th row of this matrix gives the start and end indices of the \(n\)th instance of pattern in the \(i\)th element of string.

Details

str_locate_first(...) is just str_locate_nth(..., n = 1).

str_locate_last(...) is just str_locate_nth(..., n = -1).

Examples

str_locate_nth(c("abcdabcxyz", "abcabc"), "abc", 2)

#>      start end
#> [1,]     5   7
#> [2,]     4   6