Find the \(n\)th number after the \(m\)th occurrence of a pattern.

Given a string, a pattern and natural numbers n and m, find the nth number after the mth occurrence of the pattern.

str_nth_number_after_mth(string, pattern, n, m, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_nth_number_after_first(string, pattern, n, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_nth_number_after_last(string, pattern, n, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_first_number_after_mth(string, pattern, m, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_last_number_after_mth(string, pattern, m, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_first_number_after_first(string, pattern, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_first_number_after_last(string, pattern, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_last_number_after_first(string, pattern, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

str_last_number_after_last(string, pattern, decimals = FALSE,
  leading_decimals = FALSE, negs = FALSE, leave_as_string = FALSE)

Arguments

string
A character vector.

pattern
A character vector. Pattern(s) specified like the pattern(s) in the stringr package (e.g. look at stringr::str_locate()). If this has length >1 its length must be the same as that of string.

n, m
Natural numbers.

decimals
Do you want to include the possibility of decimal numbers (TRUE) or not (FALSE, the default).

leading_decimals
Do you want to allow a leading decimal point to be the start of a number?

negs
Do you want to allow negative numbers? Note that double negatives are not handled here (see the examples).

leave_as_string
Do you want to return the number as a string (TRUE) or as numeric (FALSE, the default)?

Value

A numeric vector.

Examples

string <- c("abc1abc2abc3abc4abc5abc6abc7abc8abc9",
            "abc1def2ghi3abc4def5ghi6abc7def8ghi9")
str_nth_number_after_mth(string, "abc", 1, 3)

#> [1] 3 7

str_nth_number_after_mth(string, "abc", 2, 3)

#> [1] 4 8

str_nth_number_after_first(string, "abc", 2)

#> [1] 2 2

str_nth_number_after_last(string, "abc", -1)

#> [1] 9 9

str_first_number_after_mth(string, "abc", 2)

#> [1] 2 4

str_last_number_after_mth(string, "abc", 1)

#> [1] 9 9

str_first_number_after_first(string, "abc")

#> [1] 1 1

str_first_number_after_last(string, "abc")

#> [1] 9 7

str_last_number_after_first(string, "abc")

#> [1] 9 9

str_last_number_after_last(string, "abc")

#> [1] 9 9