Remove the quoted parts of a string.

If any parts of a string are quoted (between quotation marks), remove those parts of the string, including the quotes. Run the examples and you'll know exactly how this function works.

str_remove_quoted(string)

Arguments

string
A character vector.

Value

A character vector.

Examples

string <- "\"abc\"67a\'dk\'f"
cat(string)

#> "abc"67a'dk'f

str_remove_quoted(string)

#> [1] "67af"