Extract single elements of a string and paste them together.

This is a quick way around doing a call to str_elems() followed by a call of apply(..., paste).

str_paste_elems(string, indices, sep = "")

Arguments

string

A character vector.

indices

A vector of integerish values. Negative indexing is allowed as in stringr::str_sub().

sep

A string. The separator for pasting string elements together.

Value

A character vector.

Details

Elements that don't exist e.g. element 5 of "abc" are ignored.

See also

Examples

string <- c("abc", "def", "ghi", "vwxyz")
str_paste_elems(string, 1:2)
#> [1] "ab" "de" "gh" "vw"
str_paste_elems(string, c(1, 2, 3, 4, -1))
#> [1] "abcc"  "deff"  "ghii"  "vwxyz"
str_paste_elems("abc", c(1, 5, 55, 43, 3))
#> [1] "ac"